* misc fixes

+++ src/thread/arena.hpp
@@ -148,7 +148,7 @@ public:
     addr_t *stack_ptr = __create_stack();
     thread_attr_t attrs = micron::__thread_attr_with_stack(posix::getpid(), posix::sched_other, stack_ptr, Sz);
     threads.emplace(cpu_t<true>(), micron::move(attrs), f, micron::forward<Args>(args)...);
-    sstring<16> thread_name = "arena/" + int_to_string_stack<u32, char, 12>(threads.size());
+    sstring<16> thread_name = "arena/" + int_to_string_stack<u32, char, 12>(static_cast<u32>(threads.size()));
     thread_name[15] = 0x0;
     micthread::set_name(threads.top().thread.native_handle(), thread_name.c_str());
     return threads.top();
@@ -170,7 +170,7 @@ public:
 
     thread_attr_t attrs = micron::__thread_attr_with_stack(posix::getpid(), posix::sched_other, stack_ptr, Sz);
     threads.emplace(c, micron::move(attrs), f, micron::forward<Args>(args)...);
-    sstring<16> thread_name = "arena/" + int_to_string_stack<u32, char, 12>(threads.size());
+    sstring<16> thread_name = "arena/" + int_to_string_stack<u32, char, 12>(static_cast<u32>(threads.size()));
     thread_name[15] = 0x0;
     micthread::set_name(threads.top().thread.native_handle(), thread_name.c_str());
 
@@ -193,7 +193,7 @@ public:
 
     thread_attr_t attrs = micron::__thread_attr_with_stack(posix::getpid(), posix::sched_fifo, stack_ptr, Sz, 1);
     threads.emplace(c, micron::move(attrs), f, micron::forward<Args>(args)...);
-    sstring<16> thread_name = "arena/" + int_to_string_stack<u32, char, 12>(threads.size());
+    sstring<16> thread_name = "arena/" + int_to_string_stack<u32, char, 12>(static_cast<u32>(threads.size()));
     thread_name[15] = 0x0;
     micthread::set_name(threads.top().thread.native_handle(), thread_name.c_str());
 
@@ -237,7 +237,7 @@ public:
     addr_t *stack_ptr = __create_stack();
     thread_attr_t attrs = micron::__thread_attr_with_stack(posix::getpid(), posix::sched_other, stack_ptr, Sz);
     threads.emplace(c, micron::move(attrs), f, micron::forward<Args>(args)...);
-    sstring<16> thread_name = "arena/" + int_to_string_stack<u32, char, 12>(threads.size());
+    sstring<16> thread_name = "arena/" + int_to_string_stack<u32, char, 12>(static_cast<u32>(threads.size()));
     thread_name[15] = 0x0;
     micthread::set_name(threads.top().thread.native_handle(), thread_name.c_str());
     posix::sched_setaffinity(threads.top().thread.thread_id(), sizeof(c.get()), c.get());      // best-effort
@@ -274,7 +274,7 @@ public:
     addr_t *stack_ptr = __create_stack();
     thread_attr_t attrs = micron::__thread_attr_with_stack(posix::getpid(), posix::sched_fifo, stack_ptr, Sz, 1);
     threads.emplace(c, micron::move(attrs), f, micron::forward<Args>(args)...);
-    sstring<16> thread_name = "arena/" + int_to_string_stack<u32, char, 12>(threads.size());
+    sstring<16> thread_name = "arena/" + int_to_string_stack<u32, char, 12>(static_cast<u32>(threads.size()));
     thread_name[15] = 0x0;
     micthread::set_name(threads.top().thread.native_handle(), thread_name.c_str());
     posix::sched_setaffinity(threads.top().thread.thread_id(), sizeof(c.get()), c.get());      // best-effort
@@ -587,7 +587,7 @@ public:
     // WARNING: important, make sure no temporaries are created
     auto __cn = cpu_t<true>();
     new (&threads.mut(counter)) thread_t<Tr>{ micron::move(__cn), micron::move(attrs), f, micron::forward<Args>(args)... };
-    sstring<16> thread_name = "c/" + int_to_string_stack<u32, char, 12>(counter);
+    sstring<16> thread_name = "c/" + int_to_string_stack<u32, char, 12>(static_cast<u32>(counter));
     thread_name[15] = 0x0;
     micthread::set_name(threads.mut(counter).thread.native_handle(), thread_name.c_str());
     return threads.mut(counter++);
@@ -604,7 +604,7 @@ public:
     // WARNING: important, make sure no temporaries are created
     auto __cn = cpu_t<true>();
     new (&threads.mut(counter)) thread_t<Tr>{ micron::move(__cn), micron::move(attrs) };
-    sstring<16> thread_name = "c/" + int_to_string_stack<u32, char, 12>(counter);
+    sstring<16> thread_name = "c/" + int_to_string_stack<u32, char, 12>(static_cast<u32>(counter));
     thread_name[15] = 0x0;
     micthread::set_name(threads.mut(counter).thread.native_handle(), thread_name.c_str());
     return threads.mut(counter++);
